Add Axum server
The server allows the webpage to fetch information from Sloppy such as
balances, transactions, etc.
