+++ Cargo.toml
@@ -13,3 +13,5 @@ nwc = "0.39"
 reqwest = { version = "0.12.12", features=  [ "json", "stream" ] }
 serde_json = "1.0.139"
 futures-util = "0.3.31"
+axum = "0.8.1"
+anyhow = "1.0.96"
